feat: use sguaba
* wip: use `uom` and `sguaba` in `camera` module

* wip: make sky model use bearing from sguaba

* wip: use sguaba and uom types in ray.rs

* wip: fix pattern match

* wip: cargo fix

* wip: add frame transform to aop struct and test

* wip: remove state.rs

+++ src/error.rs
@@ -2,7 +2,7 @@ use std::fmt;
 
 #[derive(Clone, Copy, Debug, PartialEq)]
 pub enum Error {
-    OddImgDim((u32, u32)),
+    OddImgDim((u16, u16)),
     EmptyRange,
     NonFinite,
 }
